move to cc_utils
